searchable single select refactor
