+++ apps/console/src/components/pages/protected/controls/authority-card.tsx
@@ -4,15 +4,11 @@ import React, { useState } from 'react'
 import { useGetAllGroups } from '@/lib/graphql-hooks/groups'
 import { ControlDetailsFieldsFragment, Group, UpdateControlInput, UpdateSubcontrolInput } from '@repo/codegen/src/schema'
 import { Card } from '@repo/ui/cardpanel'
-import { CircleUser, CircleArrowRight, ChevronDown } from 'lucide-react'
+import { CircleUser, CircleArrowRight } from 'lucide-react'
 import { Avatar } from '@/components/shared/avatar/avatar'
-import { useFormContext, Controller } from 'react-hook-form'
 import { Option } from '@repo/ui/multiple-selector'
-import { Popover, PopoverContent, PopoverTrigger } from '@repo/ui/popover'
-import { Command, CommandEmpty, CommandGroup, CommandInput, CommandItem, CommandList } from '@repo/ui/command'
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@repo/ui/tooltip'
-import useEscapeKey from '@/hooks/useEscapeKey'
-import useClickOutsideWithPortal from '@/hooks/useClickOutsideWithPortal'
+import { SearchableSingleSelect } from '@/components/shared/searchableSingleSelect/searchable-single-select'
 
 interface AuthorityCardProps {
   controlOwner?: ControlDetailsFieldsFragment['controlOwner']
@@ -50,14 +46,7 @@ const AuthorityCard: React.FC<AuthorityCardProps> = ({ controlOwner, delegate, i
         </div>
 
         {showEditable ? (
-          <SearchableSingleSelect
-            fieldName={fieldKey}
-            options={options}
-            placeholder={`Select ${label.toLowerCase()}`}
-            onChange={(val) => handleSelect(fieldKey, val)}
-            onClose={() => setEditingField(null)}
-            autoFocus
-          />
+          <SearchableSingleSelect options={options} placeholder={`Select ${label.toLowerCase()}`} onChange={(val) => handleSelect(fieldKey, val)} onClose={() => setEditingField(null)} autoFocus />
         ) : (
           <TooltipProvider disableHoverableContent>
             <Tooltip>
@@ -93,80 +82,3 @@ const AuthorityCard: React.FC<AuthorityCardProps> = ({ controlOwner, delegate, i
 }
 
 export default AuthorityCard
-
-interface SearchableSingleSelectProps {
-  fieldName: string
-  options: Option[]
-  placeholder?: string
-  onChange?: (val: string) => void
-  onClose?: () => void
-  autoFocus?: boolean
-}
-
-export const SearchableSingleSelect: React.FC<SearchableSingleSelectProps> = ({ fieldName, options, placeholder = 'Select an option...', onChange, onClose, autoFocus }) => {
-  const { control } = useFormContext()
-  const [open, setOpen] = React.useState(true)
-
-  const triggerRef = React.useRef<HTMLDivElement>(null)
-  const popoverRef = React.useRef<HTMLDivElement>(null)
-
-  useClickOutsideWithPortal(
-    () => {
-      onClose?.()
-    },
-    {
-      refs: { triggerRef, popoverRef },
-    },
-  )
-
-  useEscapeKey(() => {
-    onClose?.()
-  })
-
-  return (
-    <Controller
-      name={fieldName}
-      control={control}
-      render={({ field }) => {
-        const selected = options.find((opt) => opt.value === field.value)
-
-        return (
-          <div className="w-[200px]" ref={triggerRef}>
-            <Popover open={open} onOpenChange={setOpen}>
-              <PopoverTrigger asChild>
-                <div className="w-full flex text-sm h-10 px-3 !py-0 justify-between border bg-input-background rounded-md items-center cursor-pointer" onClick={() => setOpen(true)}>
-                  <span className="truncate">{selected?.label || placeholder}</span>
-                  <ChevronDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-                </div>
-              </PopoverTrigger>
-              <PopoverContent ref={popoverRef} className="w-[200px] p-0 !bg-input-background border" side="bottom">
-                <Command shouldFilter autoFocus={autoFocus}>
-                  <CommandInput placeholder="Search..." />
-                  <CommandList>
-                    <CommandEmpty>No results found.</CommandEmpty>
-                    <CommandGroup>
-                      {options.map((option) => (
-                        <CommandItem
-                          key={option.value}
-                          value={option.label}
-                          onSelect={() => {
-                            field.onChange(option.value)
-                            onChange?.(option.value)
-                            setOpen(false)
-                            onClose?.()
-                          }}
-                        >
-                          {option.label}
-                        </CommandItem>
-                      ))}
-                    </CommandGroup>
-                  </CommandList>
-                </Command>
-              </PopoverContent>
-            </Popover>
-          </div>
-        )
-      }}
-    />
-  )
-}
